+++ src/routes/auth/index.ts
@@ -4,3 +4,8 @@ export * from "./logout";
 export * from "./signup";
 export * from "./all-users";
 export * from "./show-user";
+export * from "./edit-user";
+export * from "./change-password";
+export * from "./all-users-with-admins";
+export * from "./disable-user";
+export * from "./enable-user";